Made avatar nullable.
We can thus delete a user's avatar by setting it to null. Note that setting it to the empty string (at least via the REST API) doesn't work because the string representation of an avatar is expected to be either an image base64-encoded in the URL or the actual URL to an already uploaded image.